Refactor TargetSelector into Repository pattern
   This commit migrates TargetSelector to the Repository pattern following
   DDD principles, with a clear separation between domain and infrastructure
   layers.

   Changes:
   - Created PauseTargetRepository interface in domain layer
     - Defines findByHelmRelease(namespace, helmReleaseName) method
   - Implemented PauseTargetRepositoryImpl in infrastructure layer
     - Migrated all TargetSelector logic
     - namespace and helmReleaseName moved from constructor to method parameters
     - Added MAYBE comment for potential future migration of constants to domain
   - Updated Pauser to use PauseTargetRepository instead of TargetSelector
     - Store namespace and helmReleaseName as instance variables
     - Instantiate PauseTargetRepositoryImpl with APIs only
   - Migrated TargetSelectorTest to PauseTargetRepositoryImplTest
     - Updated all test method calls to use new repository interface
   - Deleted old TargetSelector class and test
   - Updated refactoring plan document
     - Added note about potential future fine-grained repositories
       (DeploymentRepository, K8sServiceRepository, etc.)
